+++ csvkit/cli.py
@@ -236,7 +236,7 @@ class CSVKitUtility:
                      '1-based numbering.')
 
         self.argparser.add_argument(
-            '-V', '--version', action='version', version='%(prog)s 1.4.0',
+            '-V', '--version', action='version', version='%(prog)s 1.5.0',
             help='Display version information and exit.')
 
     def _open_input_file(self, path, opened=False):